@@ -4,10 +4,13 @@
   "description": "A React component wrapper that restricts scrolling capabilities of everything except for its children.",
   "main": "index.js",
   "scripts": {
+    "lint": "eslint 'src/*.@(js|jsx)'",
     "clean": "rm -rf ./lib",
     "build-modules": "babel ./src --optional runtime --out-dir ./lib --source-maps",
     "build": "npm run clean && npm run build-modules",
-    "build-demo": "cd demo && webpack"
+    "build-demo": "cd demo && webpack",
+    "test": "jest --config ./.jestrc",
+    "cover": "jest --config ./.jestrc --coverage"
   },
   "keywords": [
     "React",
@@ -20,22 +23,37 @@
   "author": "Teddy Ortega <tortega128@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.7.5",
-    "babel-core": "^6.7.6",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "css-loader": "^0.23.1",
-    "node-sass": "^3.4.2",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
-    "sass-loader": "^3.2.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.0"
+    "babel-cli": "^6.23.0",
+    "babel-core": "^6.23.1",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.4.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.23.0",
+    "babel-preset-stage-0": "^6.22.0",
+    "css-loader": "^0.26.2",
+    "enzyme": "^2.7.1",
+    "enzyme-to-json": "^1.5.0",
+    "eslint": "^3.17.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.10.0",
+    "jest": "^19.0.2",
+    "node-sass": "^4.5.0",
+    "pre-commit": "^1.2.2",
+    "react": "^15.4.2",
+    "react-addons-test-utils": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "sass-loader": "^6.0.2",
+    "style-loader": "^0.13.2",
+    "webpack": "^2.2.1"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/orteth01/react-scroll-lock-component"
-  }
+  },
+  "pre-commit": [
+    "lint",
+    "test"
+  ]
 }
