@@ -1,62 +1,81 @@
 import React, { Component } from 'react';
 
 class ScrollLock extends Component {
-  componentDidMount() {
-    if (this.props.enabled) {
-      this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false)
+    constructor(props) {
+        super(props);
+        this.listenToWheelEvent = this.listenToWheelEvent.bind(this);
+        this.stopListeningToWheelEvent = this.stopListeningToWheelEvent.bind(this);
+        this.onScrollHandler = this.onScrollHandler.bind(this);
+        this.setScrollingElement = this.setScrollingElement.bind(this);
+        this.cancelScrollEvent = this.cancelScrollEvent.bind(this);
     }
-  }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.enabled) {
-      this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false)
-    } else {
-      this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false)
+    componentDidMount() {
+        if (this.props.enabled) {
+            this.listenToWheelEvent();
+        }
     }
-  }
 
-  componentWillUnmount() {
-    this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false);
-  }
-
-  setScrollingElement = (r) => {
-    this.scrollingElement = r ? r.firstChild : r;
-  };
-
-  onScrollHandler = (e) => {
-    const elem = this.scrollingElement;
-    const { scrollTop, scrollHeight, clientHeight } = elem;
-    const wheelDelta = e.deltaY;
-    const isDeltaPositive = wheelDelta > 0;
-
-    if (isDeltaPositive && wheelDelta > scrollHeight - clientHeight - scrollTop) {
-      elem.scrollTop = scrollHeight;
-      return this.cancelScrollEvent(e);
+    componentWillReceiveProps(nextProps) {
+        if (this.props.enabled !== nextProps.enabled) {
+            const fn = nextProps.enabled ? this.listenToWheelEvent : this.stopListeningToWheelEvent;
+            fn();
+        }
     }
-    else if (!isDeltaPositive && -wheelDelta > scrollTop) {
-      elem.scrollTop = 0;
-      return this.cancelScrollEvent(e);
+
+    componentWillUnmount() {
+        this.stopListeningToWheelEvent();
     }
-  };
 
-  cancelScrollEvent = (e) => {
-    e.stopImmediatePropagation();
-    e.preventDefault();
-    e.returnValue = false;
-    return false;
-  };
+    onScrollHandler(e) {
+        const elem = this.scrollingElement;
+        const { scrollTop, scrollHeight, clientHeight } = elem;
+        const wheelDelta = e.deltaY;
+        const isDeltaPositive = wheelDelta > 0;
 
-  render() {
-    return (
-      <div ref={this.setScrollingElement}>
-        {this.props.children}
-      </div>
-    );
-  }
+        let shouldCancelScroll = false;
+        if (isDeltaPositive && wheelDelta > scrollHeight - clientHeight - scrollTop) {
+            elem.scrollTop = scrollHeight;
+            shouldCancelScroll = true;
+        } else if (!isDeltaPositive && -wheelDelta > scrollTop) {
+            elem.scrollTop = 0;
+            shouldCancelScroll = true;
+        }
+
+        if (shouldCancelScroll) {
+            this.cancelScrollEvent(e);
+        }
+    }
+
+    setScrollingElement(r) {
+        this.scrollingElement = r && r.firstChild;
+    }
+
+    cancelScrollEvent(e) {
+        e.stopImmediatePropagation();
+        e.preventDefault();
+        return false;
+    }
+
+    listenToWheelEvent() {
+        this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false);
+    }
+
+    stopListeningToWheelEvent() {
+        this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false);
+    }
+
+    render() {
+        return (
+            <div ref={this.setScrollingElement}>
+                {this.props.children}
+            </div>
+        );
+    }
 }
 
 ScrollLock.defaultProps = {
-  enabled: true
+    enabled: true
 };
 
 export default ScrollLock;
