@@ -1,7 +1,7 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-  value: true
+    value: true
 });
 
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
@@ -19,82 +19,103 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var ScrollLock = function (_Component) {
-  _inherits(ScrollLock, _Component);
+    _inherits(ScrollLock, _Component);
 
-  function ScrollLock() {
-    var _ref;
+    function ScrollLock(props) {
+        _classCallCheck(this, ScrollLock);
 
-    var _temp, _this, _ret;
+        var _this = _possibleConstructorReturn(this, (ScrollLock.__proto__ || Object.getPrototypeOf(ScrollLock)).call(this, props));
 
-    _classCallCheck(this, ScrollLock);
-
-    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
+        _this.listenToWheelEvent = _this.listenToWheelEvent.bind(_this);
+        _this.stopListeningToWheelEvent = _this.stopListeningToWheelEvent.bind(_this);
+        _this.onScrollHandler = _this.onScrollHandler.bind(_this);
+        _this.setScrollingElement = _this.setScrollingElement.bind(_this);
+        _this.cancelScrollEvent = _this.cancelScrollEvent.bind(_this);
+        return _this;
     }
 
-    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = ScrollLock.__proto__ || Object.getPrototypeOf(ScrollLock)).call.apply(_ref, [this].concat(args))), _this), _this.setScrollingElement = function (r) {
-      _this.scrollingElement = r ? r.firstChild : r;
-    }, _this.onScrollHandler = function (e) {
-      var elem = _this.scrollingElement;
-      var scrollTop = elem.scrollTop,
-          scrollHeight = elem.scrollHeight,
-          clientHeight = elem.clientHeight;
+    _createClass(ScrollLock, [{
+        key: 'componentDidMount',
+        value: function componentDidMount() {
+            if (this.props.enabled) {
+                this.listenToWheelEvent();
+            }
+        }
+    }, {
+        key: 'componentWillReceiveProps',
+        value: function componentWillReceiveProps(nextProps) {
+            if (this.props.enabled !== nextProps.enabled) {
+                var fn = nextProps.enabled ? this.listenToWheelEvent : this.stopListeningToWheelEvent;
+                fn();
+            }
+        }
+    }, {
+        key: 'componentWillUnmount',
+        value: function componentWillUnmount() {
+            this.stopListeningToWheelEvent();
+        }
+    }, {
+        key: 'listenToWheelEvent',
+        value: function listenToWheelEvent() {
+            this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false);
+        }
+    }, {
+        key: 'stopListeningToWheelEvent',
+        value: function stopListeningToWheelEvent() {
+            this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false);
+        }
+    }, {
+        key: 'onScrollHandler',
+        value: function onScrollHandler(e) {
+            var elem = this.scrollingElement;
+            var scrollTop = elem.scrollTop,
+                scrollHeight = elem.scrollHeight,
+                clientHeight = elem.clientHeight;
 
-      var wheelDelta = e.deltaY;
-      var isDeltaPositive = wheelDelta > 0;
+            var wheelDelta = e.deltaY;
+            var isDeltaPositive = wheelDelta > 0;
 
-      if (isDeltaPositive && wheelDelta > scrollHeight - clientHeight - scrollTop) {
-        elem.scrollTop = scrollHeight;
-        return _this.cancelScrollEvent(e);
-      } else if (!isDeltaPositive && -wheelDelta > scrollTop) {
-        elem.scrollTop = 0;
-        return _this.cancelScrollEvent(e);
-      }
-    }, _this.cancelScrollEvent = function (e) {
-      e.stopImmediatePropagation();
-      e.preventDefault();
-      e.returnValue = false;
-      return false;
-    }, _temp), _possibleConstructorReturn(_this, _ret);
-  }
+            var shouldCancelScroll = false;
+            if (isDeltaPositive && wheelDelta > scrollHeight - clientHeight - scrollTop) {
+                elem.scrollTop = scrollHeight;
+                shouldCancelScroll = true;
+            } else if (!isDeltaPositive && -wheelDelta > scrollTop) {
+                elem.scrollTop = 0;
+                shouldCancelScroll = true;
+            }
 
-  _createClass(ScrollLock, [{
-    key: 'componentDidMount',
-    value: function componentDidMount() {
-      if (this.props.enabled) {
-        this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false);
-      }
-    }
-  }, {
-    key: 'componentWillReceiveProps',
-    value: function componentWillReceiveProps(nextProps) {
-      if (nextProps.enabled) {
-        this.scrollingElement.addEventListener('wheel', this.onScrollHandler, false);
-      } else {
-        this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false);
-      }
-    }
-  }, {
-    key: 'componentWillUnmount',
-    value: function componentWillUnmount() {
-      this.scrollingElement.removeEventListener('wheel', this.onScrollHandler, false);
-    }
-  }, {
-    key: 'render',
-    value: function render() {
-      return _react2.default.createElement(
-        'div',
-        { ref: this.setScrollingElement },
-        this.props.children
-      );
-    }
-  }]);
+            if (shouldCancelScroll) {
+                this.cancelScrollEvent(e);
+            }
+        }
+    }, {
+        key: 'setScrollingElement',
+        value: function setScrollingElement(r) {
+            this.scrollingElement = r && r.firstChild;
+        }
+    }, {
+        key: 'cancelScrollEvent',
+        value: function cancelScrollEvent(e) {
+            e.stopImmediatePropagation();
+            e.preventDefault();
+            return false;
+        }
+    }, {
+        key: 'render',
+        value: function render() {
+            return _react2.default.createElement(
+                'div',
+                { ref: this.setScrollingElement },
+                this.props.children
+            );
+        }
+    }]);
 
-  return ScrollLock;
+    return ScrollLock;
 }(_react.Component);
 
 ScrollLock.defaultProps = {
-  enabled: true
+    enabled: true
 };
 
 exports.default = ScrollLock;
